+++ comandaPSS/src/app/altas/modals/alta-producto/alta-producto.component.scss
@@ -13,7 +13,7 @@ ion-content {
 .container {
   width: 100%;
   height: auto;
-  background-color: var(--ion-color-light);
+  background-color: var(--ion-color-fondo);
   border-radius: 10px;
   ion-toolbar {
     border-radius: 10px 10px 0px 0px;
@@ -25,10 +25,10 @@ ion-content {
       cursor: pointer;
       color: var(--ion-color-primary-shade);
     }
-    ion-avatar {
+    ion-thumbnail {
       margin: 5px auto;
-      width: 120px;
-      height: 120px;
+      width: 68%;
+      height: auto;
       background: #8080808a;
     }
     ion-item {
