+++ comandaPSS/src/app/pages/clientes/mesa/mesa-routing.module.ts
@@ -6,13 +6,13 @@ import { MesaPage } from './mesa.page';
 const routes: Routes = [
   {
     path: '',
-    component: MesaPage
+    component: MesaPage,
   },
   {
     path: 'carta',
-    loadChildren: () => import('./carta/carta.module').then( m => m.CartaPageModule)
-  }
-
+    loadChildren: () =>
+      import('./carta/carta.module').then((m) => m.CartaPageModule),
+  },
 ];
 
 @NgModule({
