+++ comandaPSS/src/app/pages/cocinero/cocinero.page.html
@@ -1,14 +1,17 @@
-<utility-header type="menu" title="Comidas a preparar"></utility-header>
-<ion-content>
-  <ion-list-header>
-    <ion-label>COMIDAS A PREPARAR</ion-label>
-  </ion-list-header>
-  <ion-spinner name="circular" *ngIf="!comidasPendientes"></ion-spinner>
-
-  <div *ngIf="comidasPendientes">
-    <ion-list *ngFor="let preparacion of comidasPendientes">
-      <ion-item-sliding>
-        <ion-item>
+<utility-header
+  type="menu"
+  title="Comidas a preparar"
+  color="tertiary"
+></utility-header>
+<ion-content color="fondo" class="ion-padding">
+  <ion-list>
+    <ion-list-header>
+      <ion-label>COMIDAS A PREPARAR</ion-label>
+    </ion-list-header>
+    <ion-spinner name="circular" *ngIf="!comidasPendientes"></ion-spinner>
+    <div *ngIf="comidasPendientes">
+      <ion-item-sliding *ngFor="let preparacion of comidasPendientes">
+        <ion-item color="input">
           <ion-avatar slot="start">
             <img [src]="preparacion.producto.foto1" />
           </ion-avatar>
@@ -24,18 +27,20 @@
           ></ion-item-option>
         </ion-item-options>
       </ion-item-sliding>
-    </ion-list>
-  </div>
+      <ion-text *ngIf="comidasPendientes.length == 0">
+        <h5>No hay comidas a preparar</h5>
+      </ion-text>
+    </div>
+  </ion-list>
 
-  <ion-list-header>
-    <ion-label>COMIDAS PREPARANDOSE</ion-label>
-  </ion-list-header>
-  <ion-spinner name="circular" *ngIf="!comidasPreparando"></ion-spinner>
-
-  <div *ngIf="comidasPreparando">
-    <ion-list *ngFor="let preparacion of comidasPreparando">
-      <ion-item-sliding>
-        <ion-item>
+  <ion-list>
+    <ion-list-header>
+      <ion-label>COMIDAS PREPARANDOSE</ion-label>
+    </ion-list-header>
+    <ion-spinner name="circular" *ngIf="!comidasPreparando"></ion-spinner>
+    <div *ngIf="comidasPreparando">
+      <ion-item-sliding *ngFor="let preparacion of comidasPreparando">
+        <ion-item color="input">
           <ion-avatar slot="start">
             <img [src]="preparacion.producto.foto1" />
           </ion-avatar>
@@ -51,15 +56,18 @@
           ></ion-item-option>
         </ion-item-options>
       </ion-item-sliding>
-    </ion-list>
-  </div>
+      <ion-text *ngIf="comidasPreparando.length == 0">
+        <h5>No hay comidas en preparacion</h5>
+      </ion-text>
+    </div>
+  </ion-list>
   <!-- <ion-button (click)="debugCrearPedido()" expand="block" fill="clear" shape="round">
     Click me
   </ion-button> -->
 </ion-content>
 
 <ion-fab vertical="bottom" horizontal="end" slot="fixed">
-  <ion-fab-button (click)="presentModal()">
+  <ion-fab-button (click)="presentModal()" color="tertiary">
     <ion-icon name="add-outline"></ion-icon>
   </ion-fab-button>
 </ion-fab>
