+++ comandaPSS/src/app/pages/mozo/components/lista-chats/lista-chats.component.scss
@@ -17,3 +17,8 @@ ion-spinner {
 ion-list {
   background-color: transparent;
 }
+
+ion-item {
+  border-radius: 5px;
+  margin: 10px 0px;
+}
