+++ comandaPSS/src/app/pages/mozo/mozo.page.html
@@ -1,13 +1,12 @@
 <ion-tabs>
   <ion-tab-bar slot="bottom" color="input">
     <ion-tab-button tab="pedidos">
-      <ion-icon name="flash"></ion-icon>
+      <ion-icon name="list-outline"></ion-icon>
       <ion-label>Pedidos</ion-label>
     </ion-tab-button>
     <ion-tab-button tab="lista-chats">
-      <ion-icon name="apps"></ion-icon>
+      <ion-icon name="chatbubble-ellipses-outline"></ion-icon>
       <ion-label>Consultas</ion-label>
     </ion-tab-button>
   </ion-tab-bar>
 </ion-tabs>
-
